split out response object
