+++ Api/OrderDetailsResponse.php
@@ -1,0 +1,14 @@
+<?php
+namespace Drip\Connect\Api;
+
+class OrderDetailsResponse  extends \Magento\Framework\DataObject
+{
+    /**
+     * Get order url
+     *
+     * @return string
+     */
+    public function getOrderUrl() {
+        return $this->getData('order_url');
+    }
+}
