+++ Api/ResourceDetails.php
@@ -18,27 +18,34 @@ class ResourceDetails implements ResourceDetailsInterface
     protected $catalogProductMediaConfigFactory;
 
     /**
-    * @var \Drip\Connect\Api\ResponseFactory
+    * @var \Drip\Connect\Api\OrderDetailsResponseFactory
     */
-    protected $responseFactory;
+    protected $responseFactory1;
+
+		/**
+    * @var \Drip\Connect\Api\ProductDetailsResponseFactory
+    */
+    protected $responseFactory2;
 
     public function __construct(
         \Magento\Sales\Block\Adminhtml\Order\View\Info $orderViewInfo,
         \Magento\Catalog\Model\ProductFactory $catalogProductFactory,
         \Magento\Catalog\Model\Product\Media\ConfigFactory $catalogProductMediaConfigFactory,
-        \Drip\Connect\Api\ResponseFactory $responseFactory
+        \Drip\Connect\Api\OrderDetailsResponseFactory $responseFactory1,
+        \Drip\Connect\Api\ProductDetailsResponseFactory $responseFactory2
     ) {
         $this->orderViewInfo = $orderViewInfo;
         $this->catalogProductFactory = $catalogProductFactory;
         $this->catalogProductMediaConfigFactory = $catalogProductMediaConfigFactory;
-        $this->responseFactory = $responseFactory;
+        $this->responseFactory1 = $responseFactory1;
+        $this->responseFactory2 = $responseFactory2;
     }
 
     /**
      * {@inheritdoc}
      */
     public function orderDetails($orderId) {
-        $response = $this->responseFactory->create();
+        $response = $this->responseFactory1->create();
         $url = $this->orderViewInfo->getViewUrl($orderId);
 
         $response->setData(['order_url' => $url]);
@@ -50,7 +57,7 @@ class ResourceDetails implements ResourceDetailsInterface
      * {@inheritdoc}
      */
     public function productDetails($productId) {
-        $response = $this->responseFactory->create();
+        $response = $this->responseFactory2->create();
         $product = $this->catalogProductFactory->create()->load($productId);
         $productImage = $product->getImage();
         if (!empty($productImage)) {
