+++ Api/ResourceDetailsInterface.php
@@ -4,14 +4,14 @@ interface ResourceDetailsInterface {
     /**
      * GET for order details
      * @param string $orderId
-     * @return \Drip\Connect\Api\Response
+     * @return \Drip\Connect\Api\OrderDetailsResponse
      */
     public function orderDetails($orderId);
 
     /**
      * POST for product details
      * @param string $productId
-     * @return \Drip\Connect\Api\Response
+     * @return \Drip\Connect\Api\ProductDetailsResponse
      */
     public function productDetails($productId);
 }
