+++ Api/SettingsInterface.php
@@ -6,14 +6,14 @@ interface SettingsInterface {
      * @param string $websiteId
      * @param string $accountParam
      * @param string $integrationToken
-     * @return \Drip\Connect\Api\Response
+     * @return \Drip\Connect\Api\SettingsResponse
      */
      public function updateSettings($websiteId = 0, $accountParam, $integrationToken);
 
     /**
      * GET for integration settings API
      * @param string $websiteId]
-     * @return \Drip\Connect\Api\Response
+     * @return \Drip\Connect\Api\SettingsResponse
      */
     public function showStatus($websiteId = 0);
 }
